requirements.txt
python-dotenv==0.20.0
python_telegram_bot==20.0a0
requests==2.27.1
